Simplify files layout. Add new result observers and metric reporters.
